# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: English Anywhere Lab API
  version: 0.4.0
  description: |
    Single OpenAPI contract for English Anywhere Lab.
    Base path is /api/v1.
    This file is the only source of truth for API schema.

    Operations marked with `x-implemented: true` are live and covered by
    integration tests. Operations with `x-implemented: false` are roadmap
    placeholders — do NOT include them in contract tests or QA automation.
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a user
      operationId: register
      x-implemented: true
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      operationId: login
      x-implemented: true
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      operationId: refreshToken
      x-implemented: true
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me:
    get:
      tags: [Profile]
      summary: Get current user
      operationId: getMe
      x-implemented: true
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me/profile:
    patch:
      tags: [Profile]
      summary: Update learning profile
      operationId: updateProfile
      x-implemented: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /plans/bootstrap:
    post:
      tags: [Plan]
      summary: Bootstrap the first weekly plan
      operationId: bootstrapPlan
      x-implemented: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BootstrapPlanRequest'
      responses:
        '200':
          description: Plan generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyPlanResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /plans/today:
    get:
      tags: [Plan]
      summary: Get today's learning plan
      operationId: getTodayPlan
      x-implemented: true
      parameters:
        - in: query
          name: timezone
          schema:
            type: string
          required: false
      responses:
        '200':
          description: Today plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyPlanResponse'
  /plans/{plan_id}/tasks/{task_id}/complete:
    post:
      tags: [Plan]
      summary: Mark task completed
      operationId: completeTask
      x-implemented: true
      parameters:
        - $ref: '#/components/parameters/PlanId'
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteTaskRequest'
      responses:
        '200':
          description: Completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskCompletionResponse'
  /reviews/queue:
    get:
      tags: [Review]
      summary: Get due review cards
      operationId: getReviewQueue
      x-implemented: true
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 30
      responses:
        '200':
          description: Review queue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewQueueResponse'
  /reviews/submit:
    post:
      tags: [Review]
      summary: Submit one review result
      operationId: submitReview
      x-implemented: true
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewSubmitRequest'
      responses:
        '200':
          description: Review accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewSubmitResponse'
        '409':
          description: Duplicate idempotency key accepted as success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewSubmitResponse'
  /packs:
    get:
      tags: [Pack]
      summary: List learning packs
      operationId: listPacks
      x-implemented: true
      parameters:
        - in: query
          name: domain
          schema:
            type: string
        - in: query
          name: level
          schema:
            $ref: '#/components/schemas/CefrLevel'
        - in: query
          name: source
          schema:
            type: string
            enum: [official, ai]
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Pack list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackListResponse'
  /packs/{pack_id}:
    get:
      tags: [Pack]
      summary: Get pack detail
      operationId: getPackDetail
      x-implemented: true
      parameters:
        - $ref: '#/components/parameters/PackId'
      responses:
        '200':
          description: Pack detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /packs/{pack_id}/enroll:
    post:
      tags: [Pack]
      summary: Enroll a pack into current learning plan
      operationId: enrollPack
      x-implemented: true
      parameters:
        - $ref: '#/components/parameters/PackId'
      responses:
        '200':
          description: Enrolled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericMessage'
  /packs/generate:
    post:
      tags: [Pack]
      summary: Create AI generation job
      operationId: createPackGenerationJob
      x-implemented: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeneratePackRequest'
      responses:
        '202':
          description: Job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerationJobResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /packs/generation-jobs/{job_id}:
    get:
      tags: [Pack]
      summary: Get generation job status
      operationId: getGenerationJob
      x-implemented: true
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Job detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerationJobResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /progress/summary:
    get:
      tags: [Progress]
      summary: Get progress summary
      operationId: getProgressSummary
      x-implemented: true
      parameters:
        - in: query
          name: range
          required: true
          schema:
            type: string
            enum: [7d, 30d, 90d]
      responses:
        '200':
          description: Summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgressSummaryResponse'
  /progress/daily:
    get:
      tags: [Progress]
      summary: Get daily progress series
      operationId: getProgressDaily
      x-implemented: true
      parameters:
        - in: query
          name: from
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: to
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Daily data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgressDailyResponse'
  /progress/weekly-report:
    get:
      tags: [Progress]
      summary: Get weekly report aggregation
      operationId: getWeeklyReport
      x-implemented: true
      parameters:
        - in: query
          name: week_start
          required: true
          schema:
            type: string
            format: date
          description: ISO 8601 date of a Monday (start of the week)
      responses:
        '200':
          description: Weekly report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyReportResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /progress/monthly-report:
    get:
      tags: [Progress]
      summary: Get monthly report with skill breakdown and weakness analysis
      operationId: getMonthlyReport
      x-implemented: true
      parameters:
        - in: query
          name: month
          required: true
          schema:
            type: string
            pattern: '^\d{4}-(0[1-9]|1[0-2])$'
          description: Year-month in YYYY-MM format
      responses:
        '200':
          description: Monthly report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonthlyReportResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /lessons/{lesson_id}/output-tasks:
    get:
      tags: [Output]
      summary: List output tasks for a lesson
      operationId: listOutputTasks
      x-implemented: true
      parameters:
        - $ref: '#/components/parameters/LessonId'
      responses:
        '200':
          description: Writing tasks list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutputTaskListResponse'
  /output-tasks/{output_task_id}/submit:
    post:
      tags: [Output]
      summary: Submit a writing task for AI feedback
      operationId: submitWriting
      x-implemented: true
      parameters:
        - $ref: '#/components/parameters/OutputTaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitWritingRequest'
      responses:
        '200':
          description: Submission with AI feedback
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /output-tasks/submissions/{submission_id}:
    get:
      tags: [Output]
      summary: Get a submission by ID
      operationId: getSubmission
      x-implemented: true
      parameters:
        - $ref: '#/components/parameters/SubmissionId'
      responses:
        '200':
          description: Submission detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /sync/events:
    post:
      tags: [Sync]
      summary: Push offline events with idempotency
      operationId: pushSyncEvents
      x-implemented: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncEventsRequest'
      responses:
        '200':
          description: Ack statuses per event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncEventsResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /sync/changes:
    get:
      tags: [Sync]
      summary: Pull server-side changes since cursor
      operationId: pullSyncChanges
      x-implemented: true
      parameters:
        - in: query
          name: cursor
          required: false
          schema:
            type: string
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 200
      responses:
        '200':
          description: Server change set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncChangesResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PlanId:
      in: path
      name: plan_id
      required: true
      schema:
        type: string
        format: uuid
    TaskId:
      in: path
      name: task_id
      required: true
      schema:
        type: string
        format: uuid
    PackId:
      in: path
      name: pack_id
      required: true
      schema:
        type: string
        format: uuid
    JobId:
      in: path
      name: job_id
      required: true
      schema:
        type: string
        format: uuid
    LessonId:
      in: path
      name: lesson_id
      required: true
      schema:
        type: string
        format: uuid
    OutputTaskId:
      in: path
      name: output_task_id
      required: true
      schema:
        type: string
        format: uuid
    SubmissionId:
      in: path
      name: submission_id
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Transient server error — client should retry the entire batch with exponential backoff
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    CefrLevel:
      type: string
      enum: [A1, A2, B1, B2, C1, C2]
...
    GenericMessage:
      type: object
      properties:
        message:
          type: string
      required: [message]
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
      required: [code, message]
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        locale:
          type: string
        timezone:
          type: string
      required: [id, email, locale, timezone]
    UserLearningProfile:
      type: object
      properties:
        current_level:
          $ref: '#/components/schemas/CefrLevel'
        target_domain:
          type: string
        daily_minutes:
          type: integer
        weekly_goal_days:
          type: integer
      required: [current_level, target_domain, daily_minutes, weekly_goal_days]
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        locale:
          type: string
          default: zh-CN
        timezone:
          type: string
          default: Asia/Shanghai
      required: [email, password]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    RefreshRequest:
      type: object
      properties:
        refresh_token:
          type: string
      required: [refresh_token]
    AuthTokens:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
      required: [access_token, refresh_token, expires_in]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        tokens:
          $ref: '#/components/schemas/AuthTokens'
      required: [user, tokens]
    MeResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        learning_profile:
          $ref: '#/components/schemas/UserLearningProfile'
      required: [user, learning_profile]
    UpdateProfileRequest:
      type: object
      properties:
        current_level:
          $ref: '#/components/schemas/CefrLevel'
        target_domain:
          type: string
        daily_minutes:
          type: integer
          minimum: 5
          maximum: 180
        weekly_goal_days:
          type: integer
          minimum: 1
          maximum: 7
    BootstrapPlanRequest:
      type: object
      properties:
        level:
          $ref: '#/components/schemas/CefrLevel'
        target_domain:
          type: string
        daily_minutes:
          type: integer
          minimum: 5
          maximum: 180
        days:
          type: integer
          minimum: 3
          maximum: 14
          default: 7
      required: [level, target_domain, daily_minutes]
...
    DailyPlan:
      type: object
      properties:
        plan_id:
          type: string
          format: uuid
        date:
          type: string
          format: date
        mode:
          type: string
          enum: [fast_10m, full_20m]
        total_estimated_minutes:
          type: integer
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/PlanTask'
      required: [plan_id, date, mode, total_estimated_minutes, tasks]
    DailyPlanResponse:
      type: object
      properties:
        daily_plan:
          $ref: '#/components/schemas/DailyPlan'
      required: [daily_plan]
    WeeklyPlanResponse:
      type: object
      properties:
        week_start:
          type: string
          format: date
        daily_plans:
          type: array
          items:
            $ref: '#/components/schemas/DailyPlan'
      required: [week_start, daily_plans]
    CompleteTaskRequest:
      type: object
      properties:
        completed_at:
          type: string
          format: date-time
        duration_seconds:
          type: integer
          minimum: 0
      required: [completed_at]
    TaskCompletionResponse:
      type: object
      properties:
        task_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [completed]
        next_recommendation:
          type: string
      required: [task_id, status]
    ReviewCard:
      type: object
      properties:
        card_id:
          type: string
          format: uuid
        user_card_state_id:
          type: string
          format: uuid
        front_text:
          type: string
        back_text:
          type: string
        example_text:
          type: string
        due_at:
          type: string
          format: date-time
      required: [card_id, user_card_state_id, front_text, back_text, due_at]
    ReviewQueueResponse:
      type: object
      properties:
        due_count:
          type: integer
        cards:
          type: array
          items:
            $ref: '#/components/schemas/ReviewCard'
      required: [due_count, cards]
    ReviewSubmitRequest:
      type: object
      properties:
        card_id:
          type: string
          format: uuid
        user_card_state_id:
          type: string
          format: uuid
        rating:
          type: string
          enum: [again, hard, good, easy]
        reviewed_at:
          type: string
          format: date-time
        response_ms:
          type: integer
          minimum: 0
        client_event_id:
          type: string
          format: uuid
      required: [card_id, user_card_state_id, rating, reviewed_at, client_event_id]
    ReviewSubmitResponse:
      type: object
      properties:
        accepted:
          type: boolean
        card_id:
          type: string
          format: uuid
        next_due_at:
          type: string
          format: date-time
        scheduled_days:
          type: integer
        status:
          type: string
          enum: [learning, review, relearning]
      required: [accepted, card_id, next_due_at, scheduled_days, status]
    Pack:
      type: object
      properties:
        id:
          type: string
          format: uuid
        source:
          type: string
          enum: [official, ai]
        title:
          type: string
        description:
          type: string
        domain:
          type: string
        pack_template:
          $ref: '#/components/schemas/PackTemplate'
        level:
          $ref: '#/components/schemas/CefrLevel'
        estimated_minutes:
          type: integer
      required: [id, source, title, domain, pack_template, level, estimated_minutes]
    PackListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Pack'
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
      required: [items, page, page_size, total]
    PackDetailResponse:
      type: object
      properties:
        pack:
          $ref: '#/components/schemas/Pack'
        lessons:
          type: array
          items:
            type: object
            properties:
              lesson_id:
                type: string
                format: uuid
              title:
                type: string
              lesson_type:
                type: string
                enum: [reading, listening, speaking, writing, mixed]
              position:
                type: integer
              exercise_types:
                type: array
                items:
                  $ref: '#/components/schemas/ExerciseType'
              output_task_count:
                type: integer
            required: [lesson_id, title, lesson_type, position]
      required: [pack, lessons]
    GeneratePackRequest:
      type: object
      properties:
        level:
          $ref: '#/components/schemas/CefrLevel'
        domain:
          type: string
        pack_template:
          $ref: '#/components/schemas/PackTemplate'
        daily_minutes:
          type: integer
          minimum: 5
          maximum: 180
        days:
          type: integer
          minimum: 3
          maximum: 14
          default: 7
        focus_skills:
          type: array
          items:
            type: string
            enum: [listening, speaking, reading, writing]
      required: [level, domain, pack_template, daily_minutes]
    GenerationJobResponse:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued, running, success, failed]
        created_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
          nullable: true
        pack_id:
          type: string
          format: uuid
          nullable: true
        error_message:
          type: string
          nullable: true
      required: [job_id, status, created_at]
    ProgressSummaryResponse:
      type: object
      properties:
        range:
          type: string
        total_minutes:
          type: integer
        active_days:
          type: integer
        review_accuracy:
          type: number
          format: float
        cards_reviewed:
          type: integer
        streak_count:
          type: integer
      required: [range, total_minutes, active_days, review_accuracy, cards_reviewed, streak_count]
...
    ProgressDailyResponse:
      type: object
      properties:
        points:
          type: array
          items:
            $ref: '#/components/schemas/ProgressDailyPoint'
      required: [points]
...
    WeeklyReportResponse:
      type: object
      properties:
        week_start:
          type: string
          format: date
        total_minutes:
          type: integer
        active_days:
          type: integer
        cards_reviewed:
          type: integer
        cards_new:
          type: integer
        lessons_completed:
          type: integer
        listening_minutes:
          type: integer
        speaking_tasks:
          type: integer
        writing_tasks:
          type: integer
        streak:
          type: integer
        weekly_goal_days:
          type: integer
        goal_achieved:
          type: boolean
        review_health:
          $ref: '#/components/schemas/ReviewHealth'
        daily_breakdown:
          type: array
          items:
            $ref: '#/components/schemas/WeeklyReportDailyPoint'
        previous_week_comparison:
          $ref: '#/components/schemas/WeeklyComparison'
          nullable: true
      required: [week_start, total_minutes, active_days, cards_reviewed, cards_new, lessons_completed, listening_minutes, speaking_tasks, writing_tasks, streak, weekly_goal_days, goal_achieved, review_health, daily_breakdown]
...
    MonthlyReportResponse:
      type: object
      properties:
        month:
          type: string
          pattern: '^\d{4}-(0[1-9]|1[0-2])$'
        days_in_month:
          type: integer
        total_minutes:
          type: integer
        active_days:
          type: integer
        cards_reviewed:
          type: integer
        cards_new:
          type: integer
        lessons_completed:
          type: integer
        listening_minutes:
          type: integer
        speaking_tasks:
          type: integer
        writing_tasks:
          type: integer
        streak:
          type: integer
        monthly_goal_days:
          type: integer
        goal_achieved:
          type: boolean
        review_health:
          $ref: '#/components/schemas/ReviewHealth'
        daily_breakdown:
          type: array
          items:
            $ref: '#/components/schemas/WeeklyReportDailyPoint'
        skill_breakdown:
          $ref: '#/components/schemas/SkillBreakdown'
        weaknesses:
          type: array
          items:
            $ref: '#/components/schemas/WeaknessItem'
        previous_month_comparison:
          $ref: '#/components/schemas/MonthlyComparison'
          nullable: true
      required: [month, days_in_month, total_minutes, active_days, cards_reviewed, cards_new, lessons_completed, listening_minutes, speaking_tasks, writing_tasks, streak, monthly_goal_days, goal_achieved, review_health, daily_breakdown, skill_breakdown, weaknesses]
...
    SyncEventsRequest:
      type: object
      properties:
        events:
          type: array
          maxItems: 500
          items:
            $ref: '#/components/schemas/SyncEvent'
      required: [events]
...
    SyncEventsResponse:
      type: object
      properties:
        acks:
          type: array
          items:
            $ref: '#/components/schemas/SyncEventAck'
        server_cursor:
          type: string
      required: [acks, server_cursor]
    SyncChangesResponse:
      type: object
      properties:
        next_cursor:
          type: string
        changes:
          type: array
          items:
            type: object
            properties:
              entity_type:
                type: string
              entity_id:
                type: string
              op:
                type: string
                enum: [upsert, delete]
              payload:
                type: object
                additionalProperties: true
            required: [entity_type, entity_id, op, payload]
      required: [next_cursor, changes]
...
    OutputTaskListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OutputTask'
      required: [items]
    SubmitWritingRequest:
      type: object
      properties:
        answer_text:
          type: string
          minLength: 1
          maxLength: 5000
      required: [answer_text]
...
    SubmissionResponse:
      type: object
      properties:
        submission_id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
        answer_text:
          type: string
        feedback:
          $ref: '#/components/schemas/WritingFeedback'
        score:
          type: number
          format: float
        submitted_at:
          type: string
          format: date-time
      required: [submission_id, task_id, answer_text, score, submitted_at]